Rearrange code again.
