+++ src/solver/linear.jl
@@ -1,0 +1,1 @@
+include("linear/schur.jl")
